+++ rest/src/main/java/com/itwill/rest/repository/GroupMemberRepository.java
@@ -9,7 +9,7 @@ import com.itwill.rest.domain.GroupMemberId;
 
 public interface GroupMemberRepository extends JpaRepository<GroupMember, GroupMemberId> {
 	// 특정 아티스트의 그룹 멤버 목록을 반환하는 메서드 추가
-    List<GroupMember> findByArtist_Id(Long artistId);
+    List<GroupMember> findByArtist_Id(Integer artistId);
     
-    List<GroupMember> findByGroupId(Long groupId);
+    List<GroupMember> findByGroupId(Integer groupId);
 }
