+++ rest/src/main/java/com/itwill/rest/repository/PlayListQuerydsl.java
@@ -6,6 +6,6 @@ import com.itwill.rest.dto.PlayListFirstAlbumImgDto;
 
 public interface PlayListQuerydsl {
 
-	List<PlayListFirstAlbumImgDto> selectByUserId(Long id);
+	List<PlayListFirstAlbumImgDto> selectByUserId(Integer id);
 
 }
